Support useragent override in the constructor
When initializing the `OptableSDK` constructor, it's now possible to
override the otherwise internally assigned user-agent string, with one
specified by the caller. The caller-specified useragent is of course
optional and, if one is not assigned (or nil is passed as the useragent
at initialization time) then the default is to continue to create an
invisible WebView and extract the user-agent string from it. If a caller
specifies a non-nil useragent string, then we use it instead.

Closes https://github.com/Optable/optable-ios-sdk/issues/23

Documented in README